requirements.txt
streamlit
langchain>=0.1.0
langchain-groq
langchain-community
langchain-text-splitters
langchain-core
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
python-dotenv
